create did-document.ts helper
