+++ src/constants.ts
@@ -1,4 +1,6 @@
-export const PLACEHOLDER = '{SCID}';
+export const SCID_PLACEHOLDER = '{SCID}';
+export const DID_PLACEHOLDER = '{DID}';
+export const DID_KEY_PREFIX = 'did:key:';
 export const METHOD = 'webvh';
 export const BASE_CONTEXT = ['https://www.w3.org/ns/did/v1', 'https://w3id.org/security/multikey/v1'];
